# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a message service server with job management.",
        "title": "Message Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9005",
    "basePath": "/api",
    "paths": {
        "/job-status": {
            "post": {
                "description": "Enable or disable the message sending job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Change job status",
                "parameters": [
                    {
                        "description": "Job status (true/false)",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns job status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/messages": {
            "get": {
                "description": "Retrieves all messages that have been sent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Get all sent messages",
                "responses": {
                    "200": {
                        "description": "Returns sent messages",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/messages/{id}": {
            "get": {
                "description": "Retrieves a specific message by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Get a specific sent message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the message",
                        "schema": {
                            "$ref": "#/definitions/message.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - missing ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Message not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "message.MessageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Hello World"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "messageId": {
                    "type": "string",
                    "example": "msg123"
                },
                "sentAt": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "to": {
                    "type": "string",
                    "example": "+123545"
                }
            }
        }
    }
}